Moved EnrichKoodiUri.ts to a new folder
