+++ shared/models/AiemminHankittuPaikallinenTutkinnonOsa.ts
@@ -1,7 +1,7 @@
 import { types } from "mobx-state-tree"
 import { OsaamisenOsoittaminen } from "./OsaamisenOsoittaminen"
 import { KoodistoVastaus } from "./KoodistoVastaus"
-import { EnrichKoodiUri } from "./EnrichKoodiUri"
+import { EnrichKoodiUri } from "./Enrichment/EnrichKoodiUri"
 import { TodennettuArviointiLisatiedot } from "./TodennettuArviointiLisatiedot"
 import { AiemminHankitutTutkinnonOsatViews } from "./helpers/AiemminHankitutTutkinnonOsatViews"
 import { Organisaatio } from "./Organisaatio"
